added features & very unstable

+++ src/controller/MicroController.java
@@ -5,9 +5,9 @@ import java.util.Observer;
 
 public abstract class MicroController implements Observer, Runnable {
 	
-	private final short id;
+	private final int id;
 	
-	public MicroController(short id) {
+	public MicroController(int id) {
 		this.id = id;
 	}
